+++ internal/js/chat-stream/error_shape.js
