+++ internal/js/chat-stream/http_internal.js
