+++ internal/js/chat-stream/index.js
