+++ internal/js/chat-stream/proxy_go.js
