+++ internal/js/chat-stream/stream_emitter.js
