+++ internal/js/chat-stream/toolcall_policy.js
