+++ internal/js/chat-stream/vercel_stream.js
