+++ internal/js/helpers/stream-tool-sieve.js
