+++ internal/js/helpers/stream-tool-sieve/incremental.js
