+++ internal/js/shared/deepseek-constants.js
