+++ plans/node-syntax-gate-targets.txt
@@ -1,22 +1,22 @@
 # Node split syntax gate targets
-# Keep this list in sync with api/chat-stream and api/helpers/stream-tool-sieve split modules.
+# Keep this list in sync with api/chat-stream and internal/js/helpers/stream-tool-sieve split modules.
 
 api/chat-stream.js
-api/chat-stream/index.js
-api/chat-stream/error_shape.js
-api/chat-stream/http_internal.js
-api/chat-stream/proxy_go.js
-api/chat-stream/sse_parse.js
-api/chat-stream/stream_emitter.js
-api/chat-stream/token_usage.js
-api/chat-stream/toolcall_policy.js
-api/chat-stream/vercel_stream.js
+internal/js/chat-stream/index.js
+internal/js/chat-stream/error_shape.js
+internal/js/chat-stream/http_internal.js
+internal/js/chat-stream/proxy_go.js
+internal/js/chat-stream/sse_parse.js
+internal/js/chat-stream/stream_emitter.js
+internal/js/chat-stream/token_usage.js
+internal/js/chat-stream/toolcall_policy.js
+internal/js/chat-stream/vercel_stream.js
 
-api/helpers/stream-tool-sieve.js
-api/helpers/stream-tool-sieve/index.js
-api/helpers/stream-tool-sieve/state.js
-api/helpers/stream-tool-sieve/sieve.js
-api/helpers/stream-tool-sieve/incremental.js
-api/helpers/stream-tool-sieve/jsonscan.js
-api/helpers/stream-tool-sieve/parse.js
-api/helpers/stream-tool-sieve/format.js
+internal/js/helpers/stream-tool-sieve.js
+internal/js/helpers/stream-tool-sieve/index.js
+internal/js/helpers/stream-tool-sieve/state.js
+internal/js/helpers/stream-tool-sieve/sieve.js
+internal/js/helpers/stream-tool-sieve/incremental.js
+internal/js/helpers/stream-tool-sieve/jsonscan.js
+internal/js/helpers/stream-tool-sieve/parse.js
+internal/js/helpers/stream-tool-sieve/format.js
